proportional circle based on legend value
