+++ features/geojson/Sidebar/Legends/index.tsx
@@ -1,12 +1,20 @@
-import { LegendList } from "./LegendList";
+import { LegendList } from "./List";
 import { LegendSettings } from "./Settings";
+import { ColorSwatchs } from "./ColorSwatch";
 import { Divider } from "@components/Divider";
+import { LegendListHeader } from "./ListHeader";
+import { AssociatedKey } from "./AssociatedKey";
+import { ProportionalCircleSettings } from "./ProportionalCircle";
 
 export function LegendOptions() {
   return (
     <div>
+      <AssociatedKey />
       <LegendSettings />
-      <Divider label="Legend Value & Color Generator" />
+      <ProportionalCircleSettings />
+      <Divider />
+      <LegendListHeader />
+      <ColorSwatchs />
       <LegendList />
     </div>
   );
