+++ package.json
@@ -23,9 +23,6 @@
     "@mantine/modals": "^5.1.0",
     "@mantine/next": "^5.1.0",
     "@mantine/notifications": "^5.1.0",
-    "@turf/area": "^6.5.0",
-    "@turf/center-of-mass": "^6.5.0",
-    "@turf/centroid": "^6.5.0",
     "@turf/helpers": "^6.5.0",
     "@turf/point-on-feature": "^6.5.0",
     "chroma-js": "^2.4.2",
